Separate files for astro config
