+++ src/pages/index.astro
@@ -7,6 +7,7 @@ import "@/fonts/font-face.css";
 import "@/i18n";
 
 /**
+ * Component Script:
  * What markdown calls frontmatter, astro refers to as Component Script.
  *
  * "The code fence is designed to guarantee that the JavaScript that you
@@ -18,6 +19,13 @@ import "@/i18n";
  *  @see https://docs.astro.build/en/basics/astro-components/#the-component-script
  *  @see https://astro.new/latest/frameworks/
  *
+ * Import all `.md` files in `./src/pages/post/`:
+ * const matches = import.meta.glob('../pages/post/*.md', { eager: true });
+ *
+ * Loads and initializes the requested WASM file:
+ * const wasm = await WebAssembly.instantiateStreaming(fetch('/example.wasm'));
+ *
+ *
  * Component Props:
  * Usage: <GreetingHeadline greeting="Howdy" name="Partner" />
  *
@@ -27,6 +35,7 @@ import "@/i18n";
  * framework components that are hydrated. This means that you can only pass
  * props that don't do any HTML rendering."
  *
+ * Component Template:
  * "Use the Astro <slot /> pattern to pass static content generated by
  * Astro components as children to your framework components inside an
  * .astro component"
